Rework the CLI: Make the deploy process a subcommand
This is a bit mid right now, since there are no other commands, but it does improve the code structure a bit.
